Move stats to noting
