+++ app/src/main/java/dev/upaya/shf/ui/session/noting/stats/StatsNavigation.kt
@@ -1,4 +1,4 @@
-package dev.upaya.shf.ui.stats
+package dev.upaya.shf.ui.session.noting.stats
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -8,13 +8,13 @@ import androidx.navigation.compose.composable
 private const val routeNotingStats = "noting-stats"
 
 
-fun NavGraphBuilder.notingStats() {
+internal fun NavGraphBuilder.notingStats() {
     composable(route = routeNotingStats) {
         StatsScreen()
     }
 }
 
 
-fun NavController.navigateToNotingStats() {
+internal fun NavController.navigateToNotingStats() {
     this.navigate(route = routeNotingStats)
 }
