+++ app/src/main/java/dev/upaya/shf/ui/session/noting/stats/StatsScreen.kt
@@ -1,4 +1,4 @@
-package dev.upaya.shf.ui.stats
+package dev.upaya.shf.ui.session.noting.stats
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
@@ -7,8 +7,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import dev.upaya.shf.ui.session.SessionViewModel
-import dev.upaya.shf.ui.stats.composables.LabelStatsCard
-import dev.upaya.shf.ui.stats.composables.NotingSummaryCard
+import dev.upaya.shf.ui.session.noting.stats.composables.LabelStatsCard
+import dev.upaya.shf.ui.session.noting.stats.composables.NotingSummaryCard
 import dev.upaya.shf.ui.theme.SHFTheme
