+++ app/src/main/java/dev/upaya/shf/ui/session/noting/stats/composables/LabelFreqTable.kt
@@ -1,4 +1,4 @@
-package dev.upaya.shf.ui.stats.composables
+package dev.upaya.shf.ui.session.noting.stats.composables
 
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
