+++ app/src/main/java/dev/upaya/shf/ui/session/noting/stats/composables/NotingSummaryCard.kt
@@ -1,4 +1,4 @@
-package dev.upaya.shf.ui.stats.composables
+package dev.upaya.shf.ui.session.noting.stats.composables
 
 import android.text.format.DateUtils
 import androidx.compose.foundation.layout.*
